refactor(proposals): add proposal list components
